* Changed the endpoint 'signupDetails' to 'canSignup' which returns only boolean value
